Moved various components and utility files, including `BackButton`, `DownloadButton`, `FavoriteButton`, `FilterButton`, `FilterModal`, `Header`, `sortUtils`, and `ThemeButton`, and updated imports and references in remaining files.
